requirements.txt
celery~=5.3.6
python-dotenv~=1.0.1
